+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Listed changes are split into categories, reflecting the affected component / (s
 
 ### Archiving Core (`local_archiving`)
 
+- Display warning message when sub-plugins are configured in a way that no archiving can be performed (e.g., no storage
+  driver is enabled)
 - Display info message when no sup-plugins of a certain type are installed on manage components admin page
 - Remove alpha sub-plugins from core distribution
 - Adapt unit tests to detect optional sub-plugins